Update clients to receive and highlight a distinguished card
This must be a card that is in a player's hand to ensure that the indicator is cleared when another indicator is sent by the replay leader or because a clue was given.

The property "being in a player's hand" was refactored from HanabiClueEntry to HanabiCard as a new method.