Refactor clock message data storage to reduce redundancy and aliasing
playerTimes stores individual timer values, so only the active index needs to be stored to rebuild the game state for a redraw.

Rebuilding the UI does not directly modify the value of playerTimes. The playerTimes assignment is moved out of handleClock since the array should be assigned directly by the corresponding message from the server but not by a rebuild.

The previous implementation relied on the fact that playerTimes was an alias for the lastClock.times array. Therefore lastClock was being modified by the interval.