Ensure note indicators are set only on valid cards
If a message refers to a note for a card that is not currently drawn
according to the client replay position, store the note but do not
attempt to modify the card.

If there is a note on a played or discarded card, the note message
should not cause other clients to indicate it.