Invert child elements of cards held by others
This ensures a consistent position of visual elements on cards:

Per issue #74 arrows are positioned relative to the top of the card. The
up arrow was preserved based on font differences across OS and browser.

Per issue #78 dynamic pips are oriented toward the outside of a card
held by a player other than a user.