Update pull script to reset head to master for same file contents
Prevent changes to server files when on a branch other than master.
This is considered a failure of the script.

If the current commit hash is identical to the origin's, return.
This is considered a success since no action is needed.

If the hashes are not identical but the file contents are, reset
to match origin.

Only pull if the HEAD commit is an ancestor to the tip of origin.
This ensures that an automated merge or rebase does not leave the
server in a broken state.