feat: add Game ID to the UI during replay
This feature means that any screenshot taken during a replay will include the game ID,
so that viewers may be able to explore the context around the screenshot via replay.

This is definitely a kind of hacky approach, relying on client global total state,
rather than even just global state for a particular game. Particularly, updating the
game ID during spectator updates so that the former is updated properly when a finished
game is converted to a shared replay is a disgusting hack. However, I saw no reasonable
way to implement the feature without hacks; the server-client messaging architecture
seemed to necessitate it.