Add a docker configuration for production
The existing Dockerfile is made for a development environment. In
particular to run the docker-compose.yml file it is required to manually
compile the frontend dependencies on the host machine. Furthermore it
runs the hanabi-live daemon as root inside the docker container, which
is a security issue.

This commit adds a Dockerfile and docker-compose.yml for a production
setup in the 'docker' folder. This does not have any dependencies on the
host environment and can be built on a server having only a checkout of
this repository and the docker-compose executable. The Dockerfile uses a
common docker practice of "multi-stage builds" to achieve this. It also
mitigates most security risks by running hanabi-live as a limited user
within Docker.

A new section on how to start this docker image has been added to the
installation documentation.

Furthermore, this commit also adds a "smoketest" for this new setup to
.travis.yml. This smoketest sets up a clean environment, builds the
docker image, and then starts it. Afterwards it checks if the root URL
of the server returns HTTP 200.