feat: allow wildcard characters in the /seed endpoint
By simply using the SQL LIKE comparator in the backend,
the user can request results across multiple seeds, through
use of wildcard metacharacters.

The use case driving this feature is the NoVarathon initiative,
where we aim to collect results for seeded no variant games in
close to real time. The seed name will follow a pattern, and
we want results for all seeds matching that pattern.