Preserve password on remake table (#2822)
Closes #2683

I decided not to implement the thing that makes spectators join as
players because of Krycke's comment in #2683. (Indeed, seems wrong to
me in at least one case: if I'm say playing a game with some friends
that's pw-protected, and I restart the game, the random specs watching
that didn't enter a password probably shouldn't be joined.) So I'm gonna
say someone should open a new issue describing what the "correct"
behavior should be.