DB: Delete games upon participant deletion
This adds a trigger to the `game_participants` table so that whenever
a participant of a game is deleted
(this already automatically happens whenever a user is deleted),
the corresponding game is deleted as well.

Together, this should now ensure that for each game in the `games`
table, there is always the correct number of participants in the
`game_participants` table.
(Note that stuff can still break if games are inserted into the DB
without participants initially, or if game_participants are added
manually etc).

This fixes issue #2831.