API: Add secondary sort column id to all queries
This ensures that whenever we query data for the API,
the resulting requested sort is deterministic,
therefore yielding disjoint outputs when data is queried with
limit and offset (which is done regularly for the API to
keep answers at a reasonable size).

Note that even though it is weird, it is still defined
to query something like `ORDER BY id ASC, id DESC` in PostgreSQL,
the secondary order in that case just won't matter: The ORDER BY
clause works with lexicographic order.

See issue #2842 as well.