Client UI: Add suitLabelStackStartTexts field
This stores an array of text fields, one below each play stack,
that forms a second row similar to the ones used for the suit names
or the stack directions in Reversed/Up or Down variants.

We will use this in Sudoku variants to display the starting values of
the stacks.